feat: support to multiple pixels
