+++ package.json
@@ -28,7 +28,8 @@
     "test:watch": "vitest watch"
   },
   "dependencies": {
-    "@nuxt/kit": "^3.4.2"
+    "@nuxt/kit": "^3.4.2",
+    "minimatch": "^9.0.3"
   },
   "devDependencies": {
     "@nuxt/eslint-config": "^0.1.1",
