+++ src/types.ts
@@ -1,0 +1,12 @@
+// Module options TypeScript interface definition
+export interface ModuleOptions {
+  pixelId?: string | null;
+  track?: string;
+  autoPageView?: boolean;
+  version?: string;
+  pixels?: any[];
+  manualMode?: boolean;
+  disabled?: boolean;
+  debug?: boolean;
+  dev?: boolean;
+}
